solar detail setup
